refactoring-room-processes-image-and-added-loading-state
