+++ src/app/dining-room-process/page.tsx
@@ -1,13 +1,13 @@
-'use client';
+"use client";
 
-import DiningRoomServices from '@/components/DiningRoom';
+import DiningRoomServices from "@/components/RoomProcesses/DiningRoom";
 
 export interface PageProps {}
 
 export default function Page({}: PageProps) {
- return (
-  <div className=" p-5 md:p-7 lg:p-[42px] xl:p-20">
-   <DiningRoomServices />
-  </div>
- );
+  return (
+    <div className=" p-5 md:p-7 lg:p-[42px] xl:p-20">
+      <DiningRoomServices />
+    </div>
+  );
 }
