+++ src/app/living-room-process/page.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import LivingRoomServices from "@/components/LivingRoom";
+import LivingRoomServices from "@/components/RoomProcesses/LivingRoom";
 
 export interface PageProps {}
 
 export default function Page({}: PageProps) {
- return (
-  <div className=" p-5 md:p-7 lg:p-[42px] xl:p-20">
-   <LivingRoomServices />
-  </div>
- );
+  return (
+    <div className=" p-5 md:p-7 lg:p-[42px] xl:p-20">
+      <LivingRoomServices />
+    </div>
+  );
 }
